+++ mex/common/models/variable_group.py
@@ -13,6 +13,7 @@ from mex.common.models.base.rules import (
     PreventiveRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     ExtractedVariableGroupIdentifier,
@@ -236,12 +237,21 @@ class PreventiveVariableGroup(_Stem, PreventiveRule):
     label: list[MergedPrimarySourceIdentifier] = []
 
 
+class WorkflowVariableGroup(_Stem, WorkflowRule):
+    """Rule to prevent publishing of merged variable group items."""
+
+    entityType: Annotated[
+        Literal["WorkflowVariableGroup"], Field(alias="$type", frozen=True)
+    ] = "WorkflowVariableGroup"
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for a variable group item."""
 
     additive: AdditiveVariableGroup = AdditiveVariableGroup()
     subtractive: SubtractiveVariableGroup = SubtractiveVariableGroup()
     preventive: PreventiveVariableGroup = PreventiveVariableGroup()
+    workflow: WorkflowVariableGroup = WorkflowVariableGroup()
 
 
 class VariableGroupRuleSetRequest(_BaseRuleSet):