+++ mex/common/types/__init__.py
@@ -35,6 +35,7 @@ from mex.common.types.identifier import (
 from mex.common.types.identity import IdentityProvider
 from mex.common.types.link import Link, LinkLanguage
 from mex.common.types.path import AssetsPath, OpsPath, PathWrapper, WorkPath
+from mex.common.types.publishing_target import PublishingTarget
 from mex.common.types.sink import Sink
 from mex.common.types.temporal_entity import (
     CET,
@@ -143,6 +144,7 @@ __all__ = (
     "OpsPath",
     "PathWrapper",
     "PersonalData",
+    "PublishingTarget",
     "ResourceCreationMethod",
     "ResourceTypeGeneral",
     "Sink",