+++ mex/common/types/publishing_target.py
@@ -1,0 +1,8 @@
+from enum import Enum
+
+
+class PublishingTarget(Enum):
+    """Possible targets to which a merged item could be published."""
+
+    INVENIO = "invenio"
+    DATENKOMPASS = "datenkompass"