+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- BREAKING: New workflow rule defining forbidden targets for publishing of merged items
+- BREAKING: creation of merged items now needs publishing target parameter
+
 ### Changes
 
 ### Deprecated